finish Yelp Api data reformat and AppService data uniq
