+++ package.json
@@ -22,6 +22,7 @@
     "react-dom": "^15.4.2",
     "react-router": "^4.0.0",
     "request": "^2.81.0",
+    "underscore": "^1.8.3",
     "yelpv3": "^1.2.1"
   },
   "devDependencies": {
